refactor fetchTeams function
